requirements.txt
# Web Framework and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# LiveKit Components
livekit-agents==0.10.0
livekit-plugins-openai==0.10.1
livekit-plugins-deepgram==0.6.7
livekit-plugins-silero==0.7.1
livekit-plugins-groq>=0.1.0
livekit-plugins-cartesia>=0.1.0
livekit-api

# Knowledge Base and Vector Search
sentence-transformers>=2.2.2
qdrant-client>=1.7.0

# Web Scraping and Processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0

# Machine Learning and NLP
torch>=2.0.0
transformers>=4.36.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
pathlib>=1.0.1